Auth working properly from services
