+++ src/components/UdaSearchBox.js
@@ -3,8 +3,7 @@ import Places from './Places.js';
 import Cadastre from './Cadastre.js'
 import IconInput from './IconInput.js';
 import SearchButton from './SearchButton';
-// import { getToken } from '../services/Auth.js';
-import request from 'axios';
+import { getToken } from '../services/auth.js';
 import { searchBox, imputIconsBox } from '../stylesheets/StylesSearchBox';
 
 class UdaSearchBox extends Component {
@@ -26,28 +25,12 @@ class UdaSearchBox extends Component {
   }
 
   componentDidMount() {
-    this.getToken('adalab', '4286')
-  }
-
-  getToken(user, pwd) {
-    const reports = {
-      url: 'https://reds.urbandataanalytics.com/management/api/v1.0/login',
-      data: { 'username': user, 'password': pwd },
-      headers: { 'Content-Type': 'application/json' }
-    };
-
-    return new Promise((resolve, reject) => {
-      request.post(reports.url, reports.data, { headers: reports.headers })
-        .then(res => {
-          const authToken = res.data.authToken;
-          console.log(authToken)
-          this.setState({ token: authToken }, () => console.log('token', this.state.token));
-        })
-        .catch(e => {
-          //resolve(e.response.data.error)
-          resolve(e.response)
-        })
-    })
+    getToken('adalab', '4286')
+      .then((res) => {
+        const authToken = res.data.authToken;
+        console.log(authToken)
+        this.setState({ token: authToken }, () => console.log('token', this.state.token));
+      })
   }
 
   onChangeHandler(lat, lng) {
@@ -91,25 +74,25 @@ class UdaSearchBox extends Component {
     } = this.props.config;
     return (
       <div style={searchBox}>
-      <div style={imputIconsBox}>
-        <Places
-          placeholder={placeholderPlaces}
-          config={configPlaces}
-          onChangeHandler={this.onChangeHandler}
-        />
-        <Cadastre
-          placeholder={placeholderCadastre}
-          config={configCadastre}
-        />
-        <IconInput
-          statusPlaces={placesOn}
-          statusCadastre={cadastreOn}
-          placesActive={this.state.placesActive}
-          cadastreActive={this.state.cadastreActive}
-          onClickHandlerPlaces={this.onClickHandlerPlaces}
-          onClickHandlerCadastre={this.onClickHandlerCadastre}
-        />
-      </div>
+        <div style={imputIconsBox}>
+          <Places
+            placeholder={placeholderPlaces}
+            config={configPlaces}
+            onChangeHandler={this.onChangeHandler}
+          />
+          <Cadastre
+            placeholder={placeholderCadastre}
+            config={configCadastre}
+          />
+          <IconInput
+            statusPlaces={placesOn}
+            statusCadastre={cadastreOn}
+            placesActive={this.state.placesActive}
+            cadastreActive={this.state.cadastreActive}
+            onClickHandlerPlaces={this.onClickHandlerPlaces}
+            onClickHandlerCadastre={this.onClickHandlerCadastre}
+          />
+        </div>
         <SearchButton
           onSubmitHandler={this.onSubmitHandler}
           lat={this.state.lat}
