--- src/services/Auth.js
@@ -1,2 +1,0 @@
-import request from 'axios';
-
