+++ src/stylesheets/StylesSearchBox.js
@@ -1,10 +1,10 @@
 const searchBox = {
     // style: 'false',
-     display: 'flex',
-     flexDirection: 'row',
-     minWidth: '350px',
-     maxWidth: '800px',
-    
+    display: 'flex',
+    flexDirection: 'row',
+    minWidth: '350px',
+    maxWidth: '800px',
+
 }
 
 const algoliaStyles = {
@@ -18,14 +18,14 @@ const imputIconsBox = {
 }
 const buttonSearch = {
     flexGrow: '0',
-     width: '70px',
-     height: '35px',
-     padding: '5px',
-     border: '0px',
-     background: '#CA1C24',
-     color: 'white',
-     textTransform: 'uppercase',
-     left: '-66px'
+    width: '70px',
+    height: '35px',
+    padding: '5px',
+    border: '0px',
+    background: '#CA1C24',
+    color: 'white',
+    textTransform: 'uppercase',
+    left: '-66px'
 
 }
 const IconsBox = {
@@ -37,17 +37,17 @@ const IconsBox = {
 const placeIconBox = {
     display: 'inline-block',
     verticalAlign: 'middle',
-    paddingRight:'15px',
-    height:'18px',
-    width:'18px',
+    paddingRight: '15px',
+    height: '18px',
+    width: '18px',
 }
 
 const mapIconBox = {
     display: 'inline-block',
     verticalAlign: 'middle',
-    marginRight:'20px',
-    height:'18px',
-    width:'18px',
+    marginRight: '20px',
+    height: '18px',
+    width: '18px',
 }
 
 const inputBox = {
@@ -68,4 +68,4 @@ const inputBox = {
 
 
 
-export { algoliaStyles,searchBox, imputIconsBox, buttonSearch, placeIconBox, mapIconBox, IconsBox, inputBox };
+export { algoliaStyles, searchBox, imputIconsBox, buttonSearch, placeIconBox, mapIconBox, IconsBox, inputBox };
